virtual populate

+++ dev-data/data/import-dev-data.js
@@ -1,8 +1,9 @@
-const dotenv = require('dotenv');
 const fs = require('fs');
-const Tour = require('./../../models/tourModel');
 const mongoose = require('mongoose');
-dotenv.config({ path: `${__dirname}/../../config.env` });
+const dotenv = require('dotenv');
+const Tour = require('./../../models/tourModel');
+
+dotenv.config({ path: './config.env' });
 
 const DB = process.env.DATABASE.replace(
   '<PASSWORD>',
@@ -15,23 +16,37 @@ mongoose
     useCreateIndex: true,
     useFindAndModify: false,
   })
-  .then((con) => console.log('Database Connected'));
+  .then(() => console.log('DB connection successful!'));
 
-const tours = JSON.parse(fs.readFileSync(`${__dirname}/tours.json`, 'utf-8'));
+// READ JSON FILE
+const tours = JSON.parse(
+  fs.readFileSync(`${__dirname}/tours-simple.json`, 'utf-8')
+);
 
-const importData = async function () {
+// IMPORT DATA INTO DB
+const importData = async () => {
   try {
     await Tour.create(tours);
-    console.log('Data successfully loaded');
-    process.exit();
-  } catch (err) {}
+    console.log('Data successfully loaded!');
+  } catch (err) {
+    console.log(err);
+  }
+  process.exit();
 };
 
+// DELETE ALL DATA FROM DB
 const deleteData = async () => {
   try {
     await Tour.deleteMany();
-    process.exit();
-  } catch (err) {}
+    console.log('Data successfully deleted!');
+  } catch (err) {
+    console.log(err);
+  }
+  process.exit();
 };
-deleteData();
-// importData();
+
+if (process.argv[2] === '--import') {
+  importData();
+} else if (process.argv[2] === '--delete') {
+  deleteData();
+}
